+++ tests/testthat/test-add_rsq.R
@@ -58,3 +58,38 @@ test_that(
     expect_identical(p_rsq_std2$graphAttributes$Edges$labels,
                      p_pa_std_chk)
   })
+
+# Test a structure-only diagram
+
+mod <-
+  'f1 =~ x01 + x02 + x03
+   f2 =~ x04 + x05 + x06 + x07
+   f3 =~ x08 + x09 + x10
+   f4 =~ x11 + x12 + x13 + x14
+   f3 ~  f1 + f2
+   f4 ~  f1 + f3
+  '
+fit_sem <- lavaan::sem(mod, sem_example)
+est <- lavaan::parameterEstimates(fit_sem, rsquare = TRUE)[, c("lhs", "op", "rhs", "est", "pvalue")]
+m <- layout_matrix(f1 = c(1, 1),
+                   f2 = c(3, 1),
+                   f3 = c(2, 2),
+                   f4 = c(2, 3))
+p <- semPaths(fit_sem, whatLabels="est",
+        sizeMan = 5,
+        nCharNodes = 0, nCharEdges = 0,
+        edge.width = 0.8, node.width = 0.7,
+        edge.label.cex = 0.6,
+        mar = c(10,10,10,10),
+        layout = m,
+        structural = TRUE,
+        style = "lisrel",
+        DoNotPlot = TRUE)
+p2_std_chk <- est[(est$op == "r2") & (est$lhs %in% c("f3", "f4")), "est"]
+p2_std_chk <- paste0("R2=", formatC(p2_std_chk, 2, format = "f"))
+test_that(
+  "add_rsq: structural", {
+    p2 <- add_rsq(p, fit_sem)
+    expect_identical(p2$graphAttributes$Edges$labels[5:6],
+                     p2_std_chk)
+  })